Validation: return validation error objects instead of message strings
Instead of returning error messages e.g. "Found this, expected that", we now return both the error, the expectedNode and the actualNode. This should lead to more flexible error messaging and visualisation.

Additionally we return a validation tree based on the actual transportable supplied. This allows validation of transportable database to be better visualised.